examples/memory/requirements.txt
# Requirements for Personalized Agent Examples
#
# Core dependencies for 10xScale Agentflow + Mem0 + Cloud Qdrant integration

# 10xScale Agentflow framework (Main framework)
# agentflow>=0.3.0

# Memory management with Mem0
mem0ai>=0.1.30

# LLM API calls
litellm>=1.77.0

# Environment variable management
python-dotenv>=1.1.0

# Optional: For direct Qdrant interaction (if needed)
qdrant-client>=1.7.0

# Optional: For enhanced async operations
aiofiles>=23.0.0
